Chatbot/requirements.txt
streamlit
langchain
python-dotenv
langchain_community
unstructured
sentence_transformers
chromadb